requirements.txt
openai-whisper>=20240930
torch>=2.2.0
ffmpeg-python>=0.2.0
imageio-ffmpeg>=0.5.1
sounddevice>=0.4.7
numpy>=1.26.0
pynput>=1.7.7
pyperclip>=1.9.0
